layer refactor
